Split voting learner API tests and add new ones.
